move LookupForm into it's own component
