+++ src/App.js
@@ -12,27 +12,7 @@ function App() {
   );
 }
 
-const LookupForm = () => {
-  return (
-    <Form>
-      <label htmlFor="update-id">Lookup ID</label>
-      <Field id="update-id" name="id" type="text" />
-      <button type="submit">Find Transaction</button>
-    </Form>
-  );
-};
 
-const EnhancedLookupForm = withFormik({
-  mapPropsToValues({ id }) {
-    return {
-      id
-    };
-  },
-
-  handleSubmit(values, formikBag) {
-    return;
-  }
-})(LookupForm);
 
 const UpdateForm = ({ id }) => {
   // query for transaction somewhere...
